w3c-reports/12092014/mac/runned-tests/html/dom/documents/dom-tree-accessors/document.title-06-expected.txt


FAIL document.title and the empty string assert_equals: expected null but got Text node ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-06.html:17:16
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/dom/documents/dom-tree-accessors/document.title-06.html:10:5)
